refactor: split git module into high and low level
topic:refactor-split-git-module-into-high-and-low-level
